fix: code quality improvements across lib
- Fix double assignment in AxiosHeaders.accessor() that always overwrites
  internal state instead of preserving existing accessors
- Correct JSDoc for isArray in utils.js which incorrectly described it
  as checking for non-null objects instead of arrays
- Fix grammar error in toFormData.js JSDoc (a array -> an array)
- Rename misleading callback parameters in adapters.js (fn, value) to
  (adapter, name) to match what forEach actually passes
- Extract duplicated encodeUTF8 function from fetch.js and resolveConfig.js
  into shared lib/helpers/encodeUTF8.js
- Extract duplicated decodeURIComponentSafe function from fetch.js and
  http.js into shared lib/helpers/decodeURIComponentSafe.js
- Remove dead code: lib/helpers/deprecatedMethod.js and
  lib/env/classes/FormData.js which were never imported anywhere
