+++ lib/adapters/adapters.js
@@ -22,16 +22,16 @@ const knownAdapters = {
 };
 
 // Assign adapter names for easier debugging and identification
-utils.forEach(knownAdapters, (fn, value) => {
-  if (fn) {
+utils.forEach(knownAdapters, (adapter, name) => {
+  if (adapter) {
     try {
       // Null-proto descriptors so a polluted Object.prototype.get cannot turn
       // these data descriptors into accessor descriptors on the way in.
-      Object.defineProperty(fn, 'name', { __proto__: null, value });
+      Object.defineProperty(adapter, 'name', { __proto__: null, value: name });
     } catch (e) {
       // eslint-disable-next-line no-empty
     }
-    Object.defineProperty(fn, 'adapterName', { __proto__: null, value });
+    Object.defineProperty(adapter, 'adapterName', { __proto__: null, value: name });
   }
 });
