+++ lib/adapters/fetch.js
@@ -14,6 +14,8 @@ import settle from '../core/settle.js';
 import estimateDataURLDecodedBytes from '../helpers/estimateDataURLDecodedBytes.js';
 import { VERSION } from '../env/data.js';
 import { toByteStringHeaderObject } from '../helpers/sanitizeHeaderValue.js';
+import encodeUTF8 from '../helpers/encodeUTF8.js';
+import decodeURIComponentSafe from '../helpers/decodeURIComponentSafe.js';
 
 const DEFAULT_CHUNK_SIZE = 64 * 1024;
 
@@ -30,35 +32,6 @@ const DEFAULT_REQUEST_OPTIONS = {
 };
 
 const { isFunction } = utils;
-
-/**
- * Encode a UTF-8 string to a Latin-1 byte string for use with btoa().
- * This is a modern replacement for the deprecated unescape(encodeURIComponent(str)) pattern.
- *
- * @param {string} str The string to encode
- *
- * @returns {string} UTF-8 bytes as a Latin-1 string
- */
-const encodeUTF8 = (str) =>
-  encodeURIComponent(str).replace(/%([0-9A-F]{2})/gi, (_, hex) =>
-    String.fromCharCode(parseInt(hex, 16))
-  );
-
-// Node's WHATWG URL parser returns `username` and `password` percent-encoded.
-// Decode before composing the `auth` option so credentials such as
-// `my%40email.com:pass` are sent as `my@email.com:pass`. Falls back to the
-// original value for malformed input so a bad encoding never throws.
-const decodeURIComponentSafe = (value) => {
-  if (!utils.isString(value)) {
-    return value;
-  }
-
-  try {
-    return decodeURIComponent(value);
-  } catch (error) {
-    return value;
-  }
-};
 
 const test = (fn, ...args) => {
   try {
