+++ lib/adapters/http.js
@@ -35,6 +35,7 @@ import {
   asyncDecorator,
 } from '../helpers/progressEventReducer.js';
 import { estimateDataURLBufferAllocation } from '../helpers/estimateDataURLDecodedBytes.js';
+import decodeURIComponentSafe from '../helpers/decodeURIComponentSafe.js';
 
 const zlibOptions = {
   flush: zlib.constants.Z_SYNC_FLUSH,
@@ -172,22 +173,6 @@ function getTunnelingAgent(agentOptions, userHttpsAgent) {
 const supportedProtocols = platform.protocols.map((protocol) => {
   return protocol + ':';
 });
-
-// Node's WHATWG URL parser returns `username` and `password` percent-encoded.
-// Decode before composing the `auth` option so credentials such as
-// `my%40email.com:pass` are sent as `my@email.com:pass`. Falls back to the
-// original value for malformed input so a bad encoding never throws.
-const decodeURIComponentSafe = (value) => {
-  if (!utils.isString(value)) {
-    return value;
-  }
-
-  try {
-    return decodeURIComponent(value);
-  } catch (error) {
-    return value;
-  }
-};
 
 const flushOnFinish = (stream, [throttled, flush]) => {
   stream.on('end', flush).on('error', flush);
