+++ lib/core/AxiosHeaders.js
@@ -426,7 +426,7 @@ class AxiosHeaders {
   static accessor(header) {
     const internals =
       (this[$internals] =
-      this[$internals] =
+        this[$internals] ||
         {
           accessors: {},
         });
