+++ lib/helpers/resolveConfig.js
@@ -8,19 +8,7 @@ import mergeConfig from '../core/mergeConfig.js';
 import AxiosHeaders from '../core/AxiosHeaders.js';
 import setFormDataHeaders from '../core/setFormDataHeaders.js';
 import buildURL from './buildURL.js';
-
-/**
- * Encode a UTF-8 string to a Latin-1 byte string for use with btoa().
- * This is a modern replacement for the deprecated unescape(encodeURIComponent(str)) pattern.
- *
- * @param {string} str The string to encode
- *
- * @returns {string} UTF-8 bytes as a Latin-1 string
- */
-const encodeUTF8 = (str) =>
-  encodeURIComponent(str).replace(/%([0-9A-F]{2})/gi, (_, hex) =>
-    String.fromCharCode(parseInt(hex, 16))
-  );
+import encodeUTF8 from './encodeUTF8.js';
 
 function resolveConfig(config) {
   const newConfig = mergeConfig({}, config);
