+++ lib/helpers/toFormData.js
@@ -11,7 +11,7 @@ import PlatformBuffer from '../platform/node/classes/Buffer.js';
 export const DEFAULT_FORM_DATA_MAX_DEPTH = 100;
 
 /**
- * Determines if the given thing is a array or js object.
+ * Determines if the given thing is an array or js object.
  *
  * @param {string} thing - The object or array to be visited.
  *
