+++ lib/utils.js
@@ -194,9 +194,9 @@ const kindOfTest = (type) => {
 const typeOfTest = (type) => (thing) => typeof thing === type;
 
 /**
- * Determine if a value is a non-null object
+ * Determine if a value is an Array
  *
- * @param {Object} val The value to test
+ * @param {*} val The value to test
  *
  * @returns {boolean} True if value is an Array, otherwise false
  */
